refactor: move global test helpers methods to common
